commit 945e3b4bc345679cad92f0a896670030a411c8f8
    10089 phy_check() is bitwise, should be streetwise
commit 618d207ffd5dca94c4a4715afd5a15b0025530ea
    10495 libc should know how many FPU exceptions there are
commit 6760681825872e5ab549ae9309431453fa97e428
    10548 libumem doesn't free memory allocated by valloc or memalign
commit 10ea3ab2fe049f8d68567cf4cb2e04fda1e4645b
    10553 ipadm.1m should document hostmodel property
commit fec53dd46f43e58822d823ac73a281d6f0a1e782
    10511 loader: comconsole comc_setup should return boolean
commit 2c26b085561320ff0be162e984d17279aafda6a9
    10493 pts: cast between incompatible function types
commit 74d419caec602f7f32a3bdea419b7f3155d3eebc
    10492 ptsl: cast between incompatible function types
commit e2b861771c2991875108755a59b246b86f928eab
    10490 ptm: cast between incompatible function types
commit ff2c0cbf8b71c56c475188b2be1d96686bc645e9
    10489 mr_sas: cast between incompatible function types
commit 91f8444216cbe1f7457a80ed179210c918674c8b
    10488 nxge: cast between incompatible function types
commit 63638116bbacd9d275b00b4602da8dca12542a53
    10487 nca: cast between incompatible function types

+++ usr/src/uts/common/io/nxge/nxge_mac.c
@@ -2876,7 +2876,7 @@ fail:
 
 static int
 nxge_nlp2020_i2c_read(p_nxge_t nxgep, uint8_t ctrl_port, uint16_t address,
-	    uint16_t reg, uint8_t *data)
+    uint16_t reg, uint8_t *data)
 {
 	int  phy_dev, phy_reg;
 	uint16_t phy_data = 0;
@@ -3953,7 +3953,7 @@ nxge_tn1010_xcvr_init(p_nxge_t nxgep)
 	/*
 	 * The following 4 lines actually overwrites what ever the ndd command
 	 * has set. For example, by command
-	 * 	ndd -set /dev/nxge1 adv_autoneg_cap n (n = 0 or 1)
+	 *	ndd -set /dev/nxge1 adv_autoneg_cap n (n = 0 or 1)
 	 * we could set param_arr[param_autoneg].value to n.  However, because
 	 * here we assign constants to these parameters, whatever we set with
 	 * the "ndd -set" command will be replaced. So command
@@ -4314,8 +4314,8 @@ nxge_rx_mac_init(p_nxge_t nxgep)
 	uint8_t			portn;
 	npi_handle_t		handle;
 	npi_status_t		rs = NPI_SUCCESS;
-	uint16_t 		*addr16p;
-	uint16_t 		addr0, addr1, addr2;
+	uint16_t		*addr16p;
+	uint16_t		addr0, addr1, addr2;
 	xmac_rx_config_t	xconfig;
 	bmac_rx_config_t	bconfig;
 
@@ -4505,7 +4505,7 @@ nxge_status_t
 nxge_rx_mac_enable(p_nxge_t nxgep)
 {
 	npi_handle_t	handle;
-	uint8_t 	portn;
+	uint8_t		portn;
 	npi_status_t	rs = NPI_SUCCESS;
 	nxge_status_t	status = NXGE_OK;
 
@@ -5443,7 +5443,7 @@ fail:
 
 nxge_status_t
 nxge_mii_read(p_nxge_t nxgep, uint8_t xcvr_portn, uint8_t xcvr_reg,
-		uint16_t *value)
+    uint16_t *value)
 {
 	npi_status_t rs = NPI_SUCCESS;
 
@@ -5482,7 +5482,7 @@ fail:
 
 nxge_status_t
 nxge_mii_write(p_nxge_t nxgep, uint8_t xcvr_portn, uint8_t xcvr_reg,
-		uint16_t value)
+    uint16_t value)
 {
 	npi_status_t rs = NPI_SUCCESS;
 
@@ -5521,10 +5521,10 @@ fail:
 /*
  * Perform write to Clause45 serdes / transceiver device
  * Arguments:
- *	xcvr_portn: 	The IEEE 802.3 Clause45 PHYAD, it is the same as port
+ *	xcvr_portn:	The IEEE 802.3 Clause45 PHYAD, it is the same as port
  *			number if nxge_mdio_write is used for accessing the
  *			internal LSIL serdes. Otherwise PHYAD is different
- * 			for different platforms.
+ *			for different platforms.
  *	device:		With each PHYAD, the driver can use MDIO to control
  *			multiple devices inside the PHY, here "device" is an
  *			MMD (MDIO managable device).
@@ -5534,7 +5534,7 @@ fail:
  */
 nxge_status_t
 nxge_mdio_read(p_nxge_t nxgep, uint8_t xcvr_portn, uint8_t device,
-		uint16_t xcvr_reg, uint16_t *value)
+    uint16_t xcvr_reg, uint16_t *value)
 {
 	npi_status_t rs = NPI_SUCCESS;
 
@@ -5565,7 +5565,7 @@ fail:
 
 nxge_status_t
 nxge_mdio_write(p_nxge_t nxgep, uint8_t xcvr_portn, uint8_t device,
-		uint16_t xcvr_reg, uint16_t value)
+    uint16_t xcvr_reg, uint16_t value)
 {
 	npi_status_t rs = NPI_SUCCESS;
 
@@ -5597,7 +5597,7 @@ fail:
 
 nxge_status_t
 nxge_mii_check(p_nxge_t nxgep, mii_bmsr_t bmsr, mii_bmsr_t bmsr_ints,
-		nxge_link_state_t *link_up)
+    nxge_link_state_t *link_up)
 {
 	p_nxge_param_t	param_arr;
 	p_nxge_stats_t	statsp;
@@ -6679,8 +6679,8 @@ nxge_link_monitor(p_nxge_t nxgep, link_mon_enable_t enable)
 			 * argument to the check_link function.
 			 */
 			if (nxgep->xcvr.check_link) {
-				timerid = timeout(
-				    (fptrv_t)(nxgep->xcvr.check_link),
+				timerid = timeout((fptrv_t)(uintptr_t)
+				    (nxgep->xcvr.check_link),
 				    nxgep,
 				    drv_usectohz(LINK_MONITOR_PERIOD));
 				MUTEX_ENTER(&nxgep->poll_lock);
@@ -8105,7 +8105,7 @@ nxge_scan_ports_phy(p_nxge_t nxgep, p_nxge_hw_list_t hw_p)
 			goto error_exit;
 		}
 		break;
-	case 1: 	/* Only one clause45 port */
+	case 1:		/* Only one clause45 port */
 		switch (total_phy_fd) {	/* Number of clause22 ports */
 		case 3:
 			/*
@@ -8465,7 +8465,8 @@ nxge_is_valid_local_mac(ether_addr_st mac_addr)
 }
 
 static void
-nxge_bcm5464_link_led_off(p_nxge_t nxgep) {
+nxge_bcm5464_link_led_off(p_nxge_t nxgep)
+{
 
 	npi_status_t rs = NPI_SUCCESS;
 	uint8_t xcvr_portn;
@@ -8635,7 +8636,7 @@ nxge_get_num_of_xaui(uint32_t *port_pma_pmd_dev_id,
 
 	*num_xaui = 0;
 	if ((port_pma_pmd_dev_id[0]  == PHY_BCM8704_FAMILY &&
-	    port_pcs_dev_id[0] 	== PHY_BCM8704_FAMILY) ||
+	    port_pcs_dev_id[0] == PHY_BCM8704_FAMILY) ||
 	    (((port_pma_pmd_dev_id[0] & TN1010_DEV_ID_MASK)
 	    == TN1010_DEV_ID) &&
 	    ((port_pcs_dev_id[0] & TN1010_DEV_ID_MASK)
@@ -8748,8 +8749,8 @@ fail:
  * Teranetics TN1010 PHY chip supports both 1G and 10G modes, this function
  * figures out the speed of the PHY determined by the autonegotiation
  * process and sets the following 3 parameters,
- * 	nxgep->mac.portmode
- *     	nxgep->statsp->mac_stats.link_speed
+ *	nxgep->mac.portmode
+ *	nxgep->statsp->mac_stats.link_speed
  *	nxgep->statsp->mac_stats.xcvr_inuse
  */
 static nxge_status_t
